requirements.txt
django>=4.2,<5.0
djangorestframework>=3.14
django-cors-headers>=4.3
apscheduler>=3.10
Pillow>=10.0
qrcode>=7.4
face-recognition>=1.3
face-recognition-models  
opencv-python>=4.8
pyzbar>=0.1.9
pandas>=2.0
openpyxl>=3.1